ウィンドウサイズを取得するフックをモジュール化
