+++ components/gimmik/organisms/FixedSlider.tsx
@@ -18,21 +18,6 @@ import {
 import FixedSlide from "../molecules/FixedSlide";
 import { css } from "@emotion/react";
 
-const useWindowSize = (): number[] => {
-  const [size, setSize] = useState([0, 0]);
-  useLayoutEffect(() => {
-    const updateSize = (): void => {
-      setSize([window.innerWidth, window.innerHeight]);
-    };
-
-    window.addEventListener("resize", updateSize);
-    updateSize();
-
-    return () => window.removeEventListener("resize", updateSize);
-  }, []);
-  return size;
-};
-
 type Props = {
   slides: Array<JSX.Element>;
   isInFixedArea: boolean | undefined;
